On a fait une classe ProjetDansSacADos qui sera supprimee et remplacee par la class Objet dans le futur package sacADos. Mais c'est peut-etre une bonne source d'inspiration car elle implemente le bon protocole (pour que ca marche avec glouton) et est instanciee avec un Projet du package equipe. C'est ce qu'on veut. On n'a pas du tout fait la class SacADos.

+++ src/equipe/Projet.java
@@ -6,15 +6,15 @@ public class Projet {
 	private String description;
 	private Secteur secteur;
 	private Cout cout;
-	private double benefice;
+	private int benefice;
 	
 	
 	
-	public double getBenefice() {
+	public int getBenefice() {
 		return benefice;
 	}
 
-	public void setBenefice(double benefice) {
+	public void setBenefice(int benefice) {
 		this.benefice = benefice;
 	}
